I would call it a "fix" but there was nothing to fix
you did not use even a single function, so nothing would happen even if that compiled. "IconView" was not declared, "First.plist" should have been called "Rotaty.plist"...
Anyway, I tried to create the effect you are looking for. it is very basic, but it will give you a good place start.